FS-5062 - Moving tests against queries/application.py out of test_integration and into relevant unit test module
